add users, token login

+++ cave_back/controllers/products.js
@@ -1,17 +1,32 @@
 const productsRouter = require('express').Router()
 const Product = require('../models/product')
 const axios = require('axios')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
 
-productsRouter.get('/', (request, response) => {
-  Product.find({}).then(products => {
-    response.json(products.map(product => product.toJSON()))
-  })
+productsRouter.get('/', async (request, response) => {
+  const products = await Product.find({})//.populate('user', { name: 1, price: 1, weight: 1, id: 1 })
+  response.json(products.map(product => product.toJSON()))
 })
 
 
 productsRouter.post('/', async (request, response) => {
   const body = request.body
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    console.log('token missing or invalid')
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   if (body.name === undefined || body.name === '') {
     return response.status(400).json({ error: 'product name missing' })
@@ -31,20 +46,34 @@ productsRouter.post('/', async (request, response) => {
       price: body.price,
       weight: body.weight,
       date: new Date(),
+      user: user._id
     })
 
-    product.save().then(savedProduct => {
-      console.log({
-        product: savedProduct.toJSON(),
-        updatedExisting: false
-      })
-      response.json({
-        product: savedProduct.toJSON(),
-        updatedExisting: false
-      })
+    const savedProduct = await product.save()
+    user.products = user.products.concat(savedProduct._id)
+    await user.save()
+
+    console.log({
+      product: savedProduct.toJSON(),
+      updatedExisting: false
     })
+    response.json({
+      product: savedProduct.toJSON(),
+      updatedExisting: false
+    })
+
+    // product.save().then(savedProduct => {
+    //   console.log({
+    //     product: savedProduct.toJSON(),
+    //     updatedExisting: false
+    //   })
+    //   response.json({
+    //     product: savedProduct.toJSON(),
+    //     updatedExisting: false
+    //   })
+    // })
   } else {//product already exists so it's weight and price must be updated
-    await Product.findOneAndUpdate({ name: body.name },
+    const savedProduct = await Product.findOneAndUpdate({ name: body.name },
       {
         $inc: {
           price: body.price,
@@ -55,12 +84,19 @@ productsRouter.post('/', async (request, response) => {
       {
         new: true,
         upsert: true
-      }).then(savedProduct => {
-        response.json({
-          product: savedProduct.toJSON(),
-          updatedExisting: true
-        })
       })
+
+    response.json({
+      product: savedProduct.toJSON(),
+      updatedExisting: true
+    })
+
+    // then(savedProduct => {
+    //   response.json({
+    //     product: savedProduct.toJSON(),
+    //     updatedExisting: true
+    //   })
+    // })
   }
 })
 
@@ -76,15 +112,18 @@ productsRouter.get('/:id', (request, response, next) => {
     .catch(error => next(error))
 })
 
-productsRouter.delete('/:id', (request, response, next) => {
-  Product.findByIdAndRemove(request.params.id)
-    .then(result => {
-      response.status(204).end()
-    })
-    .catch(error => next(error))
+productsRouter.delete('/:id', async (request, response, next) => {
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  await Product.findByIdAndRemove(request.params.id)
+  response.status(204).end().catch(error => next(error))
 })
 
-productsRouter.put('/:id', (request, response, next) => {
+productsRouter.put('/:id', async (request, response, next) => {
   const body = request.body
 
   const product = {
@@ -92,11 +131,8 @@ productsRouter.put('/:id', (request, response, next) => {
     weight: body.weight,
   }
 
-  Product.findByIdAndUpdate(request.params.id, product, { new: true })
-    .then(updatedProduct => {
-      response.json(updatedProduct.toJSON())
-    })
-    .catch(error => next(error))
+  const updatedProduct = await Product.findByIdAndUpdate(request.params.id, product, { new: true })
+  response.json(updatedProduct.toJSON()).catch(error => next(error))
 })
 
 module.exports = productsRouter
